接管动画伤害机制

+++ DynamicPatcher/ExtensionHooks/GScreenExt.cs
@@ -16,7 +16,7 @@ namespace ExtensionHooks
         {
             Kratos.DrawActiveMessage();
             PrintTextManager.PrintText();
-            return (uint)0;
+            return 0;
         }
     }
 }
